Fix of pydoctor crash when creating symlink to index.html
When using pydoctor to create documentation for a single root module the symlink to index.html is created before index.html is created. This causes a crash when running pydoctor in Windows.

The creation of the symlink has been moved to a separate function that is called after index.html is created. This resolves the issue.